+++ NEWS.md
@@ -1,3 +1,9 @@
+# matlab2r 1.4.0
+
+* Added parsing of `logical()` and `string()`
+* Reformatting of `matlab2r()` listed output
+* `matlab2r()` now accepts strings of MATLAB code as input
+
 # matlab2r 1.3.0
 
 * Added URLs for home page and bug reports to DESCRIPTION